+++ crates/compiler/src/errors.rs
@@ -39,6 +39,7 @@ use crate::path::PathString;
 use crate::pitch_table::{
     InstrumentHintFreq, PlayPitchFrequency, PlayPitchSampleRate, MAX_N_PITCHES,
 };
+use crate::songs::MainVolume;
 use crate::sound_effects::MAX_SFX_TICKS;
 use crate::time::{Bpm, CommandTicks, TickClock, TickCounter, ZenLen};
 use crate::value_newtypes::{I8WithByteHexValueNewType, SignedValueNewType, UnsignedValueNewType};
@@ -218,6 +219,10 @@ pub enum ValueError {
 
     CannotConvertBpmToTickClock,
 
+    MainVolumeOutOfRange(i32),
+    MainVolumeOutOfRangeU32(u32),
+    MainVolumeHexOutOfRange(u32),
+
     EchoEdlLargerThanMaxEdl { edl: EchoEdl, max_edl: EchoEdl },
     EchoEdlOutOfRange(u32),
     EchoVolumeOutOfRange(u32),
@@ -292,6 +297,7 @@ pub enum ValueError {
     NoEchoVolume,
     NoRelativeEchoVolume,
     NoEchoFeedback,
+    NoMainVolume,
     NoFirTap,
     NoRelativeEchoFeedback,
     NoFirCoefficient,
@@ -1196,6 +1202,12 @@ impl Display for ValueError {
 
             Self::CannotConvertBpmToTickClock => write!(f, "cannot convert BPM to tick clock"),
 
+            Self::MainVolumeOutOfRange(v) => out_of_range!("main volume", v, MainVolume),
+            Self::MainVolumeOutOfRangeU32(v) => out_of_range!("main volume", v, MainVolume),
+            Self::MainVolumeHexOutOfRange(v) => {
+                write!(f, "cannot parse main volume: ${v:x} is not a byte value")
+            }
+
             Self::EchoEdlLargerThanMaxEdl { edl, max_edl } => {
                 write!(
                     f,
@@ -1333,6 +1345,7 @@ impl Display for ValueError {
             Self::NoEchoVolume => write!(f, "no echo volume value"),
             Self::NoRelativeEchoVolume => write!(f, "no relative echo volume"),
             Self::NoEchoFeedback => write!(f, "no echo feedback value"),
+            Self::NoMainVolume => write!(f, "no main volume value"),
             Self::NoFirTap => write!(f, "no fir tap"),
             Self::NoRelativeEchoFeedback => write!(f, "no relative echo feedback"),
             Self::NoFirCoefficient => write!(f, "no fir coefficient"),